change the i2c frequency
We had a problem of 'flat' areas on the ToF profile which produced peaks in the velocity estimates of KF. Saurav suggested we decrease the I2C frequency and peaks seems to have gone.

However, it is to be noted that increasing the frequency a long ago had solved the issue related to the ToF thread. Now I have to check if this change affects the thread period.