clean the names of `kf_state`
I am about to leave for my room and this commit represents an idea which I had that is yet to be implemented. The `kf_state` will be a part of the topic published by the thread to docking controller. The later will decide on whether to pefrom control or not based on this state of the KF.