add the docking controller
Jus to make sure that the value of current set-points are not crazy, I moved a paper up and down over the ToF sensor and observed the current profile. Though I am not well tuned to catch the square root relations, the results looks good to me---a green signal to proceed further with software development.